Move TaskEntry to upper level
